+++ src/main/java/com/coniverse/dangjang/domain/infrastructure/datastream/LogProducer.java
@@ -1,7 +1,7 @@
 package com.coniverse.dangjang.domain.infrastructure.datastream;
 
-import com.coniverse.dangjang.domain.log.dto.request.ClientLogRequest;
-import com.coniverse.dangjang.global.aop.log.ServerLog;
+import com.coniverse.dangjang.domain.log.dto.app.AppLog;
+import com.coniverse.dangjang.domain.log.dto.server.ServerLog;
 
 /**
  * log producer interface
@@ -11,7 +11,7 @@ import com.coniverse.dangjang.global.aop.log.ServerLog;
  * @deprecated fluentbit으로 대체
  */
 public interface LogProducer {
-	void sendMessage(ClientLogRequest message);
+	void sendMessage(AppLog message);
 
 	void sendMessage(ServerLog message);
 }
